#2 #4 #7 #8 #9 HTTPS communication, mTLS, payload encryption, full round trip from Sormas to App to Sormas
- Implements and integrats the client for the health departments.
- Client uses the SORMAS REST API to listen on tasks in SORMAS.
- Client creates data requests and push this to the public server.
- Client cyclically requests new submissions from the server, loads
them, generates the DTOs from Sormas for new ones and inserts them into
Sormas via the REST API.

- Communication takes place exclusively via HTTPS now.
- Client and server authenticate each other via mTLS. For this purpose,
a key-store and a trust-store are inserted into the apps for
development, which contain the keys and certificates. Outside of
development, these must be configured via parameters.

- The intermediate health department server has been taken out of the
system, the build and the Docker-Compose for the moment, as it is
unclear whether it is really needed.
- The client for the health departments, on the other hand, has been
added to the build (parent POM) and Docker-Compose and can be built as a
Docker container and run together with the DB and the public server.
- Turns off DataInitializerInvoker in server at the moment to avoid
irritation by the demo data. Concept for tests still has to be adapted.

- Code refactoring in public server.
- Corrections to the public API (Submission: pattern for encrypted
content and optional salt; Request: new KeyReference)

- Adds an internal API in the public server that allows a health
department client to register with its UUID, public key, and key
reference. This data is then sent to the apps with a data request for
the encryption of the data.

- Only encrypted data is accepted from the public server (pattern for
Base64 encoded data) and the client decrypts the data with the private
key matching the key reference.