#21 adapts API changes, uses HTTPS and implements encryption of payload
Communication is now only via https. However, the host check is disabled
for the dummy.

Encryption of the user data with the public key of the health department
is fully implemented. The app receives the key with the DataRequest.

In addition, the random checkcode was added, which can also simplify
development in particular.