#22 integrate the teleCode to request and submission
A teleCode unique to the health department is created with the request
and transferred to the server. Overall, this is only unique with the RKI
code of the office, which can be determined with the zip code.

In order to determine the RKI code for the zip code of an index case in
the public server, a corresponding resolution is integrated using the
data provided by the RKI.

The resolution of a teleCode to a request is implemented in the public
server.

In the dummy app, the usage of a teleCode is implemented in a
rudimentary way (use parameter -t).