#14 implements all missing functions for an MVP for events
Events delivered by the apps are now processed in the client. Events are
created in SORMAS, the relationship to the case is established, and
follow-up tasks are created.

The processing of the different data is now encapsulated in individual
processors according to the strategy pattern. The abstract base class
DataSubmissionProcessor contains the general functions like decryption
and the basic functions to create a task.

Minimum data for an event (title and description) can be entered in the
dummy app.