#11 implements all missing functions for an MVP for guests/participants
In the DataRequest there is now a field for more information about the
request. This can be used to further describe the requested data, such
as table, seat, areas, ....

For tasks that are assigned to events and are of the Investigation type,
a data request for guests/participants is now started.

Guests and participants delivered by the apps are now processed in the
client. Participants and the associated person are created in SORMAS,
the relationship to the event is established, and follow-up tasks are
created.

Minimum data for an participant (first and last name) can be entered in
the dummy app.