iris-gateway/iris-backlog#32 implements tests for public server
- DefaultApiController
- DataSubmissionHdController
- X509Authentication

For convenience, an annotation @IrisWebIntegrationTest is now available
that summarizes important annotations for integration testing.

Tests with @IrisWebIntegrationTest will use the SecurityConfiguration
which disables TLS authentication so that the tests can be performed
more easily.

DataInitializer can be implemented to write data to the database (H2) at
startup, which can then be used in tests. The DataInitializerInvoker
executes the DataInitializers.