iris-gateway/iris-backlog#27 split Docker-Compose files for targeted use
Splits the big Docker-Compose files into specific ones for subsets of
the applications with which the apps can be combined for a targeted use.
https://docs.docker.com/compose/extends/#multiple-compose-files
The files use 'extends' to avoid duplication as much as possible.
https://docs.docker.com/compose/extends/#extending-services

Now there are also separate files for staging without port opening and
local development with port opening.

Updates SORMAS versions and moves IRIS versions to .env.

Cleanup: e.g. removes HD server and unnecessary files