#20 Time values are represented uniformly as instant
We decided that we will represent time values uniformly as Instant,
because for internal values (e.g. creation time) the time zone of
creation is irrelevant and the frontend will only provide time
points/instant, which will be sufficient in the long run. Therefore we
take the simplest data type for this.