Feature/search client tests (#81)
I have written the tests for the search client.

Changes were made to findByProviderIdAndLocationId by adding null checks

The Todo should not be necessary anymore. The mapper has been tested and works.
No origin for the initial problem could be found in the iris-location-service. getLocation with providerId and locationId has no tests yet, but the mapper in toDto is the same as in search(String keyword) and as such should work.
The original error might already have been removed.