chore: implements alert service to create Slack messages and Zammad tickets
There is a JSON-RPC interface that can receive alarms. These are only
stored for the time being and the method returns immediately afterwards.

A job cyclically processes the stored alerts and sends them as HTTP
calls to the Slack or Zammad API, depending on the type. An alarm is
always for one system only. If the alarm was sent successfully, then it
is deleted from the database. This ensures that no alarms are lost if a
service is ever unavailable.

There are new mandatory properties for the two services used:
* Slack (`iris.alerts.slack.*`): `token` and `channel`.
* Zammad (`iris.alerts.zammad.*`): `apiAddress`, `token`, `environment`

These are preset in the `dev_env` profile to disable alarm sending.

Refs  iris-connect/iris-backlog#199