chore: obfuscate logged values and integrates alarm sending
Improve the validation code.

Implements obfuscation of the values and integrates this before logging
into the central validation method. Also the creation of an alert is
integrated there now.

The value to be logged with a potential attack is obfuscated except for
three characters before and after the potential attack.

Implements tests for data obfuscation in LoggingHelper.

Adds input validation tests to check the validation for potential
attacks, the logging, the data obfuscation and alerting.

The test uses slf4j-mock and can be only runing successfully if logback
isn't on the classpath.

Refs: iris-gateway/iris-backlog#145