Add a naughty demo location and naughty guest data
There is a location with naughty data now, which should be rejected by the location service. Another new location (ID a7184b09-f33e-4f0c-9986-cd031d0c0f41) returns naughty data in response to a data request.

With a property the naughty location can be enabled/disabled. This should avoid to many alerts from location service. The property is use-naughty-location and the default is false.

Get of the location for delete is now done via the ID and no longer via the index.

Co-authored-by: Jens Kutzsche <github@gebea.de>