chore: removes unnecessary client channel from BS-EPS settings
The Backend Service is only called by other services and does not itself
need to open any connections to the outside world. Therefore, the
configuration of a gRPC client is unnecessary. The existing gRPC client
causes the EPS of the BS to unnecessarily check every minute whether it
has to open a connection to the apps.