fix: international phone numbers will be accepted now
There were two errors:

* In the course of filtering dangerous data, the + at the beginning of
the phone number was also evaluated as a potential attack and the data
was sorted out. This is now corrected.

* Furthermore, the RegEx for the phone number was wrong and did not
consider numbers with more than 3 digits in the area code if separated
with spaces (e.g. +49 35204 ...). This RegEx was also very complicated,
which is unnecessary here. We now use a much simpler one.