chore: integrates feedback module and Github into alerts and renames it
The creation of tickets in Github was integrated as another possible
target next to the already existing ones and adapted to the others.

The return values for all dispatch types have been revised so that more
info and in particular the ticket number can be returned and, for
example, can be returned to the sender when providing feedback.

The central processing and creation of feedbacks was integrated as a
further variant (synchronous processing) alongside alerts (asynchronous
processing). Both were combined in the more general package `messages`.

The standalone module has been removed.

Refs iris-connect/iris-backlog#28