feat: Function built in to allow users of the IRIS front end to provide feedback directly to IRIS developers.
The JSON-RPC method for feedback requests has been added. The information of such a request is used to create a ticket in a Zammad.

The central processing and creation of feedbacks was integrated as a
further variant (synchronous processing) alongside alerts (asynchronous
processing). Both were combined in the more general package `messages`.

Additional the creation of tickets in Github was integrated as another possible
target next to the already existing ones and adapted to the others.

The return values for all dispatch types have been revised so that more
info and in particular the ticket number can be returned and, for
example, can be returned to the sender when providing feedback.

Co-authored-by: niggi0 <69158206+niggi0@users.noreply.github.com>
Refs iris-connect/iris-backlog#28
PR #217