fix: extends central configuration to use a combination of a global base and an additional profile config file
Adds configuration files for the current used stages (dev, test, prod).

Adds Spring profiles for the current used stages.

Uses Jackson updater and `@JsonMerge` to merge the JSON files.

Changes `hdProxyConfigs` from an array to a map to allow a name based merge of the configuration instead of an error-prone index based merge.