chore(build): removes build and trivy check for demo apps
The demo apps are used only by `develop` and not tagged. Therefore we can remove the building of the images and the trivy checks as they are useless.

With the introduction of DCT, the push of the images from the demo apps was also missing, which led to errors in the check and thus build.