fix(HD Search): adds `getAllHds` and improves the pervormance of the HD search with `searchForHd`
- Adds the boolean param `withDetails` to the method `searchForHd` to switch between a basic and a full data view. The basic view contains only `name`, `rkiCode` and `epsName`.
- Adds the boolean param `alsoNotConnectedHds` to the method `searchForHd` to enable if the hds without an `epsName` (these are not connected to EPS) include into the response list or not.
- Adds the method `getAllHds` to get a list of all infos about health departments from RKI data. This method has also the parameter `withDetails` and `alsoNotConnectedHds`.

When reading data from the RKI XML with XML Beam, each read costs time and the less data that needs to be read/fetched, the faster the search and processing goes. Therefore, the processing can be accelerated very significantly with the two parameters.

**It should always be loaded as little data as possible and best both parameters should be set to FALSE.

Further optimizations were made in the code to load data from the XML only once and only when it is really needed.

Refs iris-connect/iris-client#678